backend/requirements.txt
virtualenv
fastapi 
pymongo 
uvicorn
openai
python-dotenv
pydantic